+++ bidding-server/src/main/java/com/uet/bidding/model/Bid.java
@@ -1,23 +1,25 @@
 package com.uet.bidding.model;
 
-import java.io.Serializable;import java.time.LocalDateTime;
+import java.io.Serializable;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
 
 public class Bid implements Serializable {
 
   private static final long serialVersionUID = 1L;
 
   private String bidderName;
-  private double amount;
+  private BigDecimal amount;
   private LocalDateTime time;
 
-  public Bid(String bidderName, double amount) {
+  public Bid(String bidderName, BigDecimal amount) {
     this.bidderName = bidderName;
     this.amount = amount;
     this.time = LocalDateTime.now();
   }
 
   // Getter/Setter...
-  public double getAmount() {
+  public BigDecimal getAmount() {
     return amount;
   }
